implemented iterative deepening.
iterative deepening is implemented. It seems to work ok for gravityOn.
I'm not sure about gravityOff. There still seems to be some problems
though. I'm not sure if it's the heuristic or the search but the AI
seems content to ignore very strong human boards. I can make a row of
four no problem.